# Conflicts:
#	packages/direflow-component/src/DireflowComponent.tsx

+++ packages/direflow-component/src/config/config-overrides.ts
@@ -1,12 +1,11 @@
 import EventHooksPlugin from 'event-hooks-webpack-plugin';
-import HtmlWebpackExternalsPlugin from 'html-webpack-externals-plugin';
 import FilterWarningsPlugin from 'webpack-filter-warnings-plugin';
 import rimraf from 'rimraf';
 import fs from 'fs';
 import { resolve } from 'path';
 import { PromiseTask } from 'event-hooks-webpack-plugin/lib/tasks';
 
-module.exports = function override(config: any, env: any): any {
+module.exports = function override(config: any, env: string): any {
   const overridenConfig = {
     ...addWelcomeMessage(config, env),
     module: overrideModule(config.module),
@@ -18,7 +17,7 @@ module.exports = function override(config: any, env: any): any {
   return overridenConfig;
 };
 
-const addWelcomeMessage = (config: any, env: any) => {
+const addWelcomeMessage = (config: any, env: string) => {
   if (env === 'production') {
     return config;
   }
@@ -70,23 +69,34 @@ const overrideOutput = (output: any) => {
   return {
     ...newOutput,
     filename: 'componentBundle.js',
+    chunkFilename: 'vendor.js',
   };
 };
 
-const overrideOptimization = (optimization: any, env: any) => {
+const overrideOptimization = (optimization: any, env: string) => {
   const newOptions = optimization.minimizer[0].options;
+  const vendorSplitChunks = {
+    cacheGroups: {
+      vendor: {
+        test: /node_modules/,
+        chunks: 'initial',
+        name: 'vendor',
+        enforce: true,
+      }
+    }
+  };
 
   newOptions.sourceMap = env === 'development';
   optimization.minimizer[0].options = newOptions;
 
   return {
     ...optimization,
-    splitChunks: false,
+    splitChunks: shouldUseVendor(env) ? vendorSplitChunks : false,
     runtimeChunk: false,
   };
 };
 
-const overridePlugins = (plugins: any, env: any) => {
+const overridePlugins = (plugins: any, env: string) => {
   plugins[0].options.inject = 'head';
 
   plugins.push(
@@ -101,19 +111,10 @@ const overridePlugins = (plugins: any, env: any) => {
     }),
   );
 
-  const customScripts = getCustomScripts();
-  if (customScripts) {
-    plugins.push(
-      new HtmlWebpackExternalsPlugin({
-        externals: customScripts,
-      }),
-    );
-  }
-
   return plugins;
 };
 
-const copyBundleScript = async (env: any) => {
+const copyBundleScript = async (env: string) => {
   if (env !== 'production') {
     return;
   }
@@ -123,21 +124,24 @@ const copyBundleScript = async (env: any) => {
   }
 
   fs.readdirSync('build').forEach((file: string) => {
-    if (file !== 'componentBundle.js') {
+    if (file !== 'componentBundle.js' && file !== 'vendor.js') {
       rimraf.sync(`build/${file}`);
     }
   });
 };
 
-const getCustomScripts = () => {
-  const { getDireflowPlugin } = require('./dist/utils/direflowConfigExtrator');
-  const plugin = getDireflowPlugin('script-loader');
-
-  if (!plugin) {
-    return;
+const shouldUseVendor = (env: string) => {
+  if (env !== 'production') {
+    return false;
   }
 
-  if (plugin.options.externals && plugin.options.externals.length) {
-    return plugin.options.externals;
+  if (process.argv.length < 3) {
+    return false;
   }
+
+  if (!process.argv.some((arg: string) => arg === '--vendor')) {
+    return false;
+  }
+
+  return true;
 };
